Update libfdt and support current ndless-sdk versions

+++ libfdt/fdt_ro.c
@@ -494,7 +494,7 @@ int fdt_node_offset_by_phandle(const void *fdt, uint32_t phandle)
 {
 	int offset;
 
-	if ((phandle == 0) || (phandle == (uint32_t)-1))
+	if ((phandle == 0) || (phandle == -1))
 		return -FDT_ERR_BADPHANDLE;
 
 	FDT_CHECK_HEADER(fdt);
